PredicateCompiler
