+++ crates/core/src/pattern_compiler/any_compiler.rs
@@ -1,5 +1,13 @@
-use super::{compiler::CompilationContext, node_compiler::NodeCompiler};
-use crate::pattern::{any::Any, patterns::Pattern, variable::VariableSourceLocations};
+use super::{
+    compiler::CompilationContext, node_compiler::NodeCompiler,
+    predicate_compiler::PredicateCompiler,
+};
+use crate::pattern::{
+    any::{Any, PrAny},
+    patterns::Pattern,
+    predicates::Predicate,
+    variable::VariableSourceLocations,
+};
 use anyhow::Result;
 use marzano_util::analysis_logs::AnalysisLogs;
 use std::collections::BTreeMap;
@@ -40,6 +48,44 @@ impl NodeCompiler for AnyCompiler {
             Ok(patterns.remove(0))
         } else {
             Ok(Pattern::Any(Box::new(Any::new(patterns))))
+        }
+    }
+}
+
+pub(crate) struct PrAnyCompiler;
+
+impl NodeCompiler for PrAnyCompiler {
+    type TargetPattern = Predicate;
+
+    fn from_node(
+        node: &Node,
+        context: &CompilationContext,
+        vars: &mut BTreeMap<String, usize>,
+        vars_array: &mut Vec<Vec<VariableSourceLocations>>,
+        scope_index: usize,
+        global_vars: &mut BTreeMap<String, usize>,
+        logs: &mut AnalysisLogs,
+    ) -> Result<Self::TargetPattern> {
+        let mut cursor = node.walk();
+        let children = node
+            .children_by_field_name("predicates", &mut cursor)
+            .filter(|n| n.is_named());
+        let mut predicates = Vec::new();
+        for predicate in children {
+            predicates.push(PredicateCompiler::from_node(
+                &predicate,
+                context,
+                vars,
+                vars_array,
+                scope_index,
+                global_vars,
+                logs,
+            )?);
+        }
+        if predicates.len() == 1 {
+            Ok(predicates.remove(0))
+        } else {
+            Ok(Predicate::Any(Box::new(PrAny::new(predicates))))
         }
     }
 }
